data_pipeline/requirements.txt
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
dotenv==0.9.9
google-auth==2.48.0
google-auth-oauthlib==1.3.0
gspread==6.2.1
idna==3.11
iniconfig==2.3.0
numpy==2.4.2
oauthlib==3.3.1
packaging==26.0
pandas==3.0.1
pluggy==1.6.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tzdata==2025.3
urllib3==2.6.3
